requirements.txt
numpy
geopandas
shapely
pyproj
requests
Pillow
opencv-python
torch
trimesh
rasterio
scipy
matplotlib
